Refactoring all functionality out of CAN.py #99
