+++ can/__init__.py
@@ -10,7 +10,12 @@ rc = dict(channel=0)
 class CanError(IOError):
     pass
 
-from can.CAN import BufferedReader, Listener, Logger, Printer, CSVWriter, SqliteWriter, ASCWriter, LogReader, SqlReader, set_logging_level
+from can.listener import Listener, BufferedReader, RedirectReader
+from can.io import Logger, Printer, CSVWriter, SqliteWriter, ASCWriter, LogReader, SqlReader
+
+from can.util import set_logging_level
+
+
 from can.message import Message
 from can.bus import BusABC
 from can.notifier import Notifier
